requirements.txt
fastapi[all]~=0.116.1
uvicorn
argon2-cffi
sqlalchemy~=2.0.42
pydantic[email]~=2.11.7
mangum
passlib[bcrypt]~=1.7.4
passlib[argon2]
python-jose~=3.5.0
pymysql
jinja2==3.1.6
alembic~=1.16.4
PyJWT~=2.10.1
aiomysql
requests~=2.32.5
httpx==0.27.2
starlette~=0.47.2
python-dotenv~=1.1.1
cloudinary~=1.44.1
resend>=2.0.0